Gates abstraction
